fix room delete foreign keys

+++ models/Room.php
@@ -119,6 +119,11 @@ class Room
 
     public static function deleteByID(int $roomId) : bool
     {
+        $stmt = PDOProvider::get()->prepare("SELECT * FROM `employee` WHERE room = :roomId");
+        $stmt->execute(['roomId'=>$roomId]);
+        if($stmt->fetch()){
+            return false;
+        }
         $query = "DELETE FROM `".self::DB_TABLE."` WHERE `room_id` = :roomId";
         $stmt = PDOProvider::get()->prepare($query);
         return $stmt->execute(['roomId'=>$roomId]);
